This scripts loads any *.py script in a dockable window inside Maya. This way, you can call scripts without sourcing them or restarting Maya, and they are nicely arranged in a dedicated window.
You can automatically assign an icon to the newly added script by making one and moving it inside prefs/icons/<specialFolder>.
Note: <specialFolder> can be modified to fit your needs. Refer to the variable 'self.scanFolder' at the beginning of the script.

Both the script and the icon have to respect a certain naming convention, as following:
-> category_scriptname.py (script) or .png (icon)

Inside Maya, the category will automatically be created as a frameLayout (if it does not exist yet), and a button with the name of your script (and your icon)will be added under it.

It is critical that both the script AND the icon have the same name in order for the icon to be correctly associated with the script in Maya.